I want to describe how to refactor the space:
1.
previously, we use two direction array, d[i][j]
but when we want to calculate d[i][j], we compare d[i-1][j]
and d[i-1][j-v[i-1]], so we finally can get the d[i][j]'s value, and
we don't need to save the d[i-1][j]'s value. (we do this when we use two
direction array)

2.
so, the thing is firstly we have a for loop (for[i:1-n]), everytime we
get into the i's for loop we have the previous f[i] array, because we
haven't changed it yet.

so we then need to compare f[v] with f[j-v[i-1]], remember again, the
f[j] now is the previous result, so if we compare from the end, we can
use f[j-v[i-1]] properly. and when we get the result of f[v], we can
erase the f[j-v[i-1]] without any side effects. (cool)